拆分文件逻辑
